==> 04_transformacion_vars_01.Rmd <==
---
title: "Transformación de variables en Base01"
subtitle: 
output: html_notebook
---

```{r}
library(dplyr)
library(tidyr)
library(ggplot2)
```

```{r}
base01<-readRDS("data/base01.rds")
head(base01%>%
       select(COD_CLIENTE,anio_cot,aniomes_cot,COD_SOL,one_of(names(.)[!names(.)%in%c("COD_CLIENTE","aniomes_cot","anio_cot","COD_SOL")])))
```

# Variables `factor` to `spread`
```{r}
vars_factor_spread<-c("GARANTIA","SEXO","ESTADO_CIVIL","REGION")
vars_factor_spread
```

## Distribuciones
```{r}
base01%>%
  group_by_(.dots = vars_factor_spread)%>%
  summarise(n=n())
```

### `r vars_factor_spread[1]`
```{r}
base01%>%
  group_by_(.dots = vars_factor_spread[1])%>%
  summarise(n=n())
```

### `r vars_factor_spread[2]`
```{r}
base01%>%
  group_by_(.dots = vars_factor_spread[2])%>%
  summarise(n=n())
```

### `r vars_factor_spread[3]`
```{r}
base01%>%
  group_by_(.dots = vars_factor_spread[3])%>%
  summarise(n=n())
```

### `r vars_factor_spread[4]`
```{r}
base01%>%
  group_by_(.dots = vars_factor_spread[4])%>%
  summarise(n=n())
```

## Contando `NA`s en esos campos
```{r}
sum(is.na(
  as.data.frame(
    base01%>%
      select(one_of(vars_factor_spread))
  )
))
  
```

## Contando `vacíos`s en esos campos
```{r}
sum(
  as.data.frame(
    base01%>%
      select(one_of(vars_factor_spread))
  )==""
)
  
```

## Convirtiendo a `character`
```{r}
base01<-base01%>%
  mutate_at(vars(vars_factor_spread),as.character)
```

## Pasando `vacíos` a valor especial `"-556677"`

```{r}
base02<-base01%>%
  mutate_at(
    vars(vars_factor_spread),
    function(x){
      x<-if_else(x=="","-556677",x)
      return(x)
    }
  )


sum(
  as.data.frame(
    base01%>%
      select(one_of(vars_factor_spread))
  )==""
);
sum(
  as.data.frame(
    base02%>%
      select(one_of(vars_factor_spread))
  )=="-556677"
)
```


**Vars numéricas de `Base01`**
```{r}
nombres_base00_01<-readRDS("data/vars_to_numeric_base00_01.rds")
nombres_base00_01[!grepl("RNG",nombres_base00_01)]
```


```{r}

```




## `spread`

```{r}
base02%>%
  group_by(COD_CLIENTE,aniomes_cot)%>%
  summarise(
    num_sols=n()
  )%>%
  arrange(desc(num_sols))
```
```{r}
base02%>%
  group_by(COD_CLIENTE)%>%
  summarise(
    num=n(),
    num_fhs_cot=length(unique(aniomes_cot)),
    tipos_garantia=length(unique(GARANTIA))
  )%>%
  arrange(desc(tipos_garantia),desc(num))
```

```{r}
base02a<-base02%>%
  select(-one_of(vars_factor_spread))

print("---vars_factor_spread")
  for(idvar00 in 1:length(vars_factor_spread)){
    # idvar00<-1
    var00<-vars_factor_spread[idvar00]
    print(paste0(">>>>>>>>>>",var00))
    
    base02a_tmp<-base02%>%
      ungroup()%>%
      # filter(COD_CLIENTE%in%c("==BBVAPE185727711"))%>%
      # mutate(var_tmp=)
      mutate_(.dots = setNames(var00, "var_tmp"))%>%
      select(
        one_of(
        c("COD_CLIENTE","aniomes_cot","COD_SOL")
        ),one_of(var00),var_tmp)%>%
      mutate(
        nombre_var_tmp=paste0(vars_factor_spread[idvar00],"_",var_tmp),
        I_vartmp=1
        # nombre_var_tmp=paste0(var00,"_",var_tmp)
        )%>%
      spread(nombre_var_tmp,I_vartmp,fill=-778899)%>%#,var_tmp,fill=-778899)%>%
      select(-one_of(var00))%>%
      group_by(COD_CLIENTE,aniomes_cot,COD_SOL)%>%
      summarise_at(vars(starts_with(var00)),max)
    
    # print(sort(paste0(var00,"_",unlist(unique(base4$tiempo_meses_cot_dat)))))
    
    base02a<-base02a%>%
      left_join(
        base02a_tmp
      )
    
    # saveRDS(catalogo_tmp03_vartmp,paste0("data/ciclos_ctlgos03_",aniomes_tmp,"_vartmp_",var00,".rds"))
    
    # length(unique(catalogo_tmp03_vartmp$CONTRATO))
    
  }

# length(vars_factor_spread)*6+ncol(base4a); ncol(base4b)
# ((length(vars_factor_spread)*6)+ncol(base4a)) == ncol(base4b)
saveRDS(base02a,"data/base02a.rds")
head(base02a)
```

```{r}
suma_categorias<-0
for(i in 1:length(vars_factor_spread)){
  # i<-1
  var_tmp<-vars_factor_spread[i]
  print(var_tmp)
  # suma_categorias<-suma_categorias+length(unique(as.data.frame(base02)[,var_tmp]))
  suma_categorias<-suma_categorias+nrow(unique(base02[,var_tmp]))
}
suma_categorias
```
```{r}
(ncol(base02)-length(vars_factor_spread)+suma_categorias); ncol(base02a)
```


# Variables a nivel Cliente + Mes de Cotización
```{r}
base02a%>%
  group_by(COD_SOL)%>%
  summarise(n=n())%>%
  group_by(n)%>%
  summarise(frew_n=n())
```

```{r}
base02a%>%
  group_by(COD_CLIENTE,aniomes_cot)%>%
  summarise(
    n_COD_SOL=length(unique(COD_SOL)),
    n=n()
    )%>%
  group_by(n_COD_SOL)%>%
  summarise(freq_n=n())
```

```{r}
base02a_cte_cot<-base02a%>%
  group_by(COD_CLIENTE,aniomes_cot)%>%
  arrange(COD_CLIENTE,aniomes_cot,FLG_DESEMBOLSO) # que `FLG_DESEMBOLSO==1` vaya a lo último de cada mes de cotización

```

```{r}

base02a_cte_cot_01<-base02a_cte_cot%>%
  mutate(
    id_solicitud_mes=row_number(),
    max_importe_mes_acum=cummax(IMPORTE),
    min_importe_mes_acum=cummin(IMPORTE),
    max_plazo_mes_acum=cummax(PLAZO),
    min_plazo_mes_acum=cummin(PLAZO),
    max_tea_mes_acum=cummax(TEA_MINIMA),
    min_tea_mes_acum=cummin(TEA_MINIMA),
    
    I_ultima_solicitud_mes=if_else(id_solicitud_mes==max(id_solicitud_mes),1,0),  ## O O O J O O O    C O N    E S T A    V A R I A B L E
    
    I_importe_menor_vs_maxacum_del_mes=if_else(IMPORTE<max_importe_mes_acum,1,0),
    I_plazo_menor_vs_maxacum_del_mes=if_else(PLAZO<max_plazo_mes_acum,1,0),
    I_tea_menor_vs_maxacum_del_mes=if_else(TEA_MINIMA<max_tea_mes_acum,1,0),
    
    I_importe_mayor_vs_minacum_del_mes=if_else(IMPORTE>min_importe_mes_acum,1,0),
    I_plazo_mayor_vs_minacum_del_mes=if_else(PLAZO>min_plazo_mes_acum,1,0),
    I_tea_mayor_vs_minacum_del_mes=if_else(TEA_MINIMA>min_tea_mes_acum,1,0),
    
    I_importe_igual_al_maxacum_del_mes=if_else(IMPORTE==max_importe_mes_acum,1,0),
    I_plazo_igual_al_maxacum_del_mes=if_else(PLAZO==max_plazo_mes_acum,1,0),
    I_tea_igual_al_maxacum_del_mes=if_else(TEA_MINIMA==max_tea_mes_acum,1,0),
    
    I_importe_igual_vs_minacum_del_mes=if_else(IMPORTE==min_importe_mes_acum,1,0),
    I_plazo_igual_al_minacum_del_mes=if_else(PLAZO==min_plazo_mes_acum,1,0),
    I_tea_igual_al_minacum_del_mes=if_else(TEA_MINIMA==min_tea_mes_acum,1,0)
  )

base02a_cte_cot_01%>%
  select(COD_CLIENTE,aniomes_cot,COD_SOL,id_solicitud_mes,I_ultima_solicitud_mes,IMPORTE,PLAZO,TEA_MINIMA,ends_with("_del_mes"),ends_with("_acum"))
```

```{r}
# saveRDS(base02a_cte_cot_01,"data/base02a_cte_cot_01.rds")
```

# Variables a nivel Cliente + Año de Cotización
```{r}
base02a_cte_aniocot<-base02a_cte_cot_01%>%
  summarise(
    anio_cot=unique(anio_cot),
    mes_cot=unique(mes_cot)
  )%>%
  ungroup()%>%
  group_by(COD_CLIENTE,anio_cot)%>%
  arrange(COD_CLIENTE,anio_cot,mes_cot)%>%
  mutate(
    id_mes_del_anio=row_number(),
    I_ultimo_mes_del_anio=if_else(id_mes_del_anio==max(id_mes_del_anio),1,0) ## O O O J O O O    C O N    E S T A    V A R I A B L E
  )%>%
  ungroup()%>%
  select(-anio_cot,-mes_cot)%>%
  right_join(
    base02a_cte_cot_01%>%
      ungroup()
  )%>%
  group_by(COD_CLIENTE,anio_cot)%>%
  arrange(COD_CLIENTE,anio_cot,mes_cot,FLG_DESEMBOLSO) # que `FLG_DESEMBOLSO==1` vaya a lo último de cada mes de cotización
head(base02a_cte_aniocot%>%
       select(COD_CLIENTE,aniomes_cot,starts_with("id_"),starts_with("I_"),ends_with("_acum"),ends_with("_del_mes"),
              one_of(names(.)[!names(.)%in%c(
                "COD_CLIENTE","aniomes_cot",
                starts_with("id_"),starts_with("I_"),ends_with("_acum"),ends_with("_del_mes"))]
                )
              ))
```

```{r}
# saveRDS(base02a_cte_aniocot,"data/base02a_cte_aniocot.rds")
```

# Variables a nivel Cliente

## Distribución de número de solicitudes por mes por Cliente
```{r}
base01%>%
  group_by(COD_CLIENTE,aniomes_cot)%>%
  summarise(num_sol=n())%>%
  ungroup()%>%
  group_by(num_sol)%>%
  summarise(freq_numsol=n())%>%
  mutate(prop=round(freq_numsol/sum(freq_numsol),4))%>%
  arrange(desc(num_sol))
```

- Los casos de más de 4 solicitudes por mes son menos del *3%* en total y donde, cada caso, representa menos del *1%*.

## Distribución de número meses por Cliente
```{r}
base01%>%
  group_by(COD_CLIENTE)%>%
  summarise(num_meses=length(unique(aniomes_cot)))%>%
  ungroup()%>%
  group_by(num_meses)%>%
  summarise(freq_num_meses=n())%>%
  mutate(prop=round(freq_num_meses/sum(freq_num_meses),4))%>%
  arrange(desc(num_meses))
```

- Los casos de más de 3 meses con solicitudes son menos del *2%* en total y donde, cada caso, representa menos del *1%*.

## Distribución de número de solicitudes totales por Cliente
```{r}
base01%>%
  group_by(COD_CLIENTE)%>%
  summarise(num_sol=n())%>%
  ungroup()%>%
  group_by(num_sol)%>%
  summarise(freq_numsol=n())%>%
  mutate(prop=round(freq_numsol/sum(freq_numsol),4))%>%
  arrange(desc(num_sol))
```

- Los casos de más de 4 solicitudes en total son menos del *3%* en total y donde, cada caso, representa menos del *1%*.

- Por lo tanto, viendo las últimas 4 solicitudes del cliente previas a la actual, sería más que suficiente

## Variables nivel Cliente
```{r}
base02a_cte<-base02a_cte_aniocot%>%
  ungroup()%>%
  group_by(COD_CLIENTE)%>%
  arrange(COD_CLIENTE,anio_cot,mes_cot,FLG_DESEMBOLSO)%>% # que `FLG_DESEMBOLSO==1` vaya a lo último de cada mes de cotización)%>%
  mutate(
    id_solicitud_del_cte=row_number(),
    I_ult_solicitud_del_cte=if_else(id_solicitud_del_cte==max(id_solicitud_del_cte),1,0),
    
    max_importe_cte_acum=cummax(IMPORTE),
    min_importe_cte_acum=cummin(IMPORTE),
    max_plazo_cte_acum=cummax(PLAZO),
    min_plazo_cte_acum=cummin(PLAZO),
    max_tea_cte_acum=cummax(TEA_MINIMA),
    min_tea_cte_acum=cummin(TEA_MINIMA),
    
    I_importe_menor_vs_maxacum_del_cte=if_else(IMPORTE<max_importe_cte_acum,1,0),
    I_plazo_menor_vs_maxacum_del_cte=if_else(PLAZO<max_plazo_cte_acum,1,0),
    I_tea_menor_vs_maxacum_del_cte=if_else(TEA_MINIMA<max_tea_cte_acum,1,0),
    
    I_importe_mayor_vs_minacum_del_cte=if_else(IMPORTE>min_importe_cte_acum,1,0),
    I_plazo_mayor_vs_minacum_del_cte=if_else(PLAZO>min_plazo_cte_acum,1,0),
    I_tea_mayor_vs_minacum_del_cte=if_else(TEA_MINIMA>min_tea_cte_acum,1,0),
    
    I_importe_igual_al_maxacum_del_cte=if_else(IMPORTE==max_importe_cte_acum,1,0),
    I_plazo_igual_al_maxacum_del_cte=if_else(PLAZO==max_plazo_cte_acum,1,0),
    I_tea_igual_al_maxacum_del_cte=if_else(TEA_MINIMA==max_tea_cte_acum,1,0),
    
    I_importe_igual_vs_minacum_del_cte=if_else(IMPORTE==min_importe_cte_acum,1,0),
    I_plazo_igual_al_minacum_del_cte=if_else(PLAZO==min_plazo_cte_acum,1,0),
    I_tea_igual_al_minacum_del_cte=if_else(TEA_MINIMA==min_tea_cte_acum,1,0),
    
    # max_importe_cte=max(IMPORTE),
    # min_importe_cte=min(IMPORTE),
    # max_plazo_cte=max(PLAZO),
    # min_plazo_cte=min(PLAZO),
    # max_tea_cte=max(TEA_MINIMA),
    # min_tea_cte=min(TEA_MINIMA)
    
    # num_exitos=cumsum()
    
    IMPORTE_m1=lag(IMPORTE,n = 1,default = -667788),
    PLAZO_m1=lag(PLAZO,n = 1,default = -667788),
    TEA_MINIMA_m1=lag(TEA_MINIMA,n = 1,default = -667788),
    
    IMPORTE_m2=lag(IMPORTE,n = 2,default = -667788),
    PLAZO_m2=lag(PLAZO,n = 2,default = -667788),
    TEA_MINIMA_m2=lag(TEA_MINIMA,n = 2,default = -667788),
    
    IMPORTE_m3=lag(IMPORTE,n = 3,default = -667788),
    PLAZO_m3=lag(PLAZO,n = 3,default = -667788),
    TEA_MINIMA_m3=lag(TEA_MINIMA,n = 3,default = -667788),
    
    IMPORTE_m4=lag(IMPORTE,n = 4,default = -667788),
    PLAZO_m4=lag(PLAZO,n = 4,default = -667788),
    TEA_MINIMA_m4=lag(TEA_MINIMA,n = 4,default = -667788),
    
    I_baja_importe_vs_m1=if_else(IMPORTE_m1!=-667788 & IMPORTE<IMPORTE_m1,1,0),
    I_baja_plazo_vs_m1=if_else(PLAZO_m1!=-667788 & PLAZO<PLAZO_m1,1,0),
    I_baja_tasa_vs_m1=if_else(TEA_MINIMA_m1!=-667788 & TEA_MINIMA<TEA_MINIMA_m1,1,0),
    
    I_baja_importe_vs_m2=if_else(IMPORTE_m2!=-667788 & IMPORTE<IMPORTE_m2,1,0),
    I_baja_plazo_vs_m2=if_else(PLAZO_m2!=-667788 & PLAZO<PLAZO_m2,1,0),
    I_baja_tasa_vs_m2=if_else(TEA_MINIMA_m2!=-667788 & TEA_MINIMA<TEA_MINIMA_m2,1,0),
    
    I_baja_importe_vs_m3=if_else(IMPORTE_m3!=-667788 & IMPORTE<IMPORTE_m3,1,0),
    I_baja_plazo_vs_m3=if_else(PLAZO_m3!=-667788 & PLAZO<PLAZO_m3,1,0),
    I_baja_tasa_vs_m3=if_else(TEA_MINIMA_m3!=-667788 & TEA_MINIMA<TEA_MINIMA_m3,1,0),
    
    I_baja_importe_vs_m4=if_else(IMPORTE_m4!=-667788 & IMPORTE<IMPORTE_m4,1,0),
    I_baja_plazo_vs_m4=if_else(PLAZO_m4!=-667788 & PLAZO<PLAZO_m4,1,0),
    I_baja_tasa_vs_m4=if_else(TEA_MINIMA_m4!=-667788 & TEA_MINIMA<TEA_MINIMA_m4,1,0),
    
    I_sube_importe_vs_m1=if_else(IMPORTE_m1!=-667788 & IMPORTE>IMPORTE_m1,1,0),
    I_sube_plazo_vs_m1=if_else(PLAZO_m1!=-667788 & PLAZO>PLAZO_m1,1,0),
    I_sube_tasa_vs_m1=if_else(TEA_MINIMA_m1!=-667788 & TEA_MINIMA>TEA_MINIMA_m1,1,0),
    
    I_sube_importe_vs_m2=if_else(IMPORTE_m2!=-667788 & IMPORTE>IMPORTE_m2,1,0),
    I_sube_plazo_vs_m2=if_else(PLAZO_m2!=-667788 & PLAZO>PLAZO_m2,1,0),
    I_sube_tasa_vs_m2=if_else(TEA_MINIMA_m2!=-667788 & TEA_MINIMA>TEA_MINIMA_m2,1,0),
    
    I_sube_importe_vs_m3=if_else(IMPORTE_m3!=-667788 & IMPORTE>IMPORTE_m3,1,0),
    I_sube_plazo_vs_m3=if_else(PLAZO_m3!=-667788 & PLAZO>PLAZO_m3,1,0),
    I_sube_tasa_vs_m3=if_else(TEA_MINIMA_m3!=-667788 & TEA_MINIMA>TEA_MINIMA_m3,1,0),
    
    I_sube_importe_vs_m4=if_else(IMPORTE_m4!=-667788 & IMPORTE>IMPORTE_m4,1,0),
    I_sube_plazo_vs_m4=if_else(PLAZO_m4!=-667788 & PLAZO>PLAZO_m4,1,0),
    I_sube_tasa_vs_m4=if_else(TEA_MINIMA_m4!=-667788 & TEA_MINIMA>TEA_MINIMA_m4,1,0)
    
    
  )
head(base02a_cte%>%
       select(COD_CLIENTE,aniomes_cot,starts_with("id_"),starts_with("I_"),ends_with("_acum"),ends_with("_del_mes"),
              one_of(names(.)[!names(.)%in%c(
                "COD_CLIENTE","aniomes_cot",
                starts_with("id_"),starts_with("I_"),ends_with("_acum"),ends_with("_del_mes"))]
                )
              ))
```

```{r}
base03<-base02a_cte%>%ungroup()

names(base03)<-gsub("-","__",names(base03))
# saveRDS(base03,"data/base03.rds")
```



```{r}
# saveRDS(base02a_cte,"data/base02a_cte.rds")
# base03<-base02a_cte%>%ungroup()
# saveRDS(base03,"data/base03.rds")
```

```{r}
base03_train<-base03%>%
  filter(muestra=="train")
saveRDS(base03_train,"data/base03_train.rds")

base03_test<-base03%>%
  filter(muestra=="test")
saveRDS(base03_test,"data/base03_test.rds")
```
